Handle fractional counts; add data refresh
Treat recipe/task counts as doubles (GraphQL floats) to avoid deserialization errors for fractional resource drains. Add a tarkov.dev market data refresh flow to EftDataManager: RefreshState, RefreshStatus/RefreshError, TriggerRefresh(), async RefreshFromApiAsync() and a refresh guard to prevent overlapping updates; background startup update now calls TriggerRefresh. Update HideoutPanel to show an empty-state UI for missing craft data with a "Download now" action, add FormatCount for fractional display, and round price calculations when multiplying by fractional counts. Format counts in QuestPlanner as well. Move local map controls and Escape "Close All" handling into the ImGui-driven RadarWindow.ImGui code (added CloseAllPanels()), and simplify RadarWindow.Input.OnKeyDown to only handle the global F8 debug toggle (other local shortcuts moved to the ImGui handler).

+++ src-silk/Misc/Data/CraftElement.cs
@@ -45,8 +45,13 @@ namespace eft_dma_radar.Silk.Misc.Data
             [JsonPropertyName("item")]
             public ItemRef? Item { get; set; }
 
+            /// <summary>
+            /// GraphQL Float — fractional for drained resources, e.g. the Water
+            /// Collector's purified-water craft uses 0.66 of a Water filter.
+            /// An int here breaks deserialization of the whole payload.
+            /// </summary>
             [JsonPropertyName("count")]
-            public int Count { get; set; }
+            public double Count { get; set; }
 
             [JsonPropertyName("attributes")]
             public List<AttributeElement>? Attributes { get; set; }
